refactor(lotto): MVC 패키지에 맞춰 클래스 분리
