+++ docs/README.md
@@ -60,8 +60,16 @@
 - [x] 리팩토링 하고 당첨번호, 보너스 번호에도 소수점 처리하자
 - [ ] 중복된 메소드가 있는지 확인
 
-- [ ] enum안에 map을 키우는 메소드를 만들 수 있는지 확인해보자
 - [ ] MVC 도입하자. 클래스가 많으니 빠르게 찾기 힘들다.
+    - [x] 일단 집어넣기
+    - [ ] view 수정
+    - [ ] controller 수정
+    - [ ] model 수정
+    - [ ] Model은 Controller와 View에 의존하지 않아야 한다.
+    - [ ] View에는 Model에만 의존해야 하고 Controller에는 의존하면 안 된다
+    - [ ] View가 Model로 부터 데이터를 받을 때는, 사용자마다 다르게 보여주어야 하는 데이터에 대해서만 받아야 한다.
+    - [ ] Controller는 Model과 View에 의존해도 된다
+    - [ ] View 가 Model로부터 데이터를 받을 때, 반드시 Controller에서 받아야 한다.
 
 
 - [ ]  이름을 통해 의도를 드러냈는가
