+++ src/main/java/lotto/Controller/AmountProcessing.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.Controller;
 
 public class AmountProcessing {
     private static final int DIVIDED = 1000;
