+++ src/main/java/lotto/Controller/CalculateLotto.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.Controller;
 
 import static lotto.MatchNums.*;
