+++ src/main/java/lotto/Controller/ProcessLotto.java
@@ -1,7 +1,12 @@
-package lotto;
+package lotto.Controller;
 
 import java.util.ArrayList;
 import java.util.List;
+import lotto.View.InputView;
+import lotto.View.OutputView;
+import lotto.Model.Generator;
+import lotto.Model.Lotto;
+import lotto.Model.StaisticsResult;
 
 public class ProcessLotto {
     private OutputView outputView;
