+++ src/test/java/lotto/GeneratorTest.java
@@ -4,6 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.HashSet;
 import java.util.List;
+import lotto.Model.Generator;
 import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
