+++ src/test/java/lotto/LottoTest.java
@@ -1,8 +1,6 @@
 package lotto;
 
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
-import org.junit.jupiter.api.Assertions;
+import lotto.Model.Lotto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
